Suppress `input` event for partial date ranges
Previously when working with a `date_field` or `datetime_field` with
`mode: :range`, the underlying `<input>` element would dispatch an
`input` event when both the start and end dates were chosen. This caused
containing forms to be submitted twice: once with just the start date,
and again with both the start and end date.

This change prevents range date/time fields with `mode: :range` from
generating an `input` event before both ends of the range have been
selected.

Date/time fields that are not operating with `mode: :range` continue to
dispatch events as normal.