fix(text-field): Fix `behavior` attribute
The `TextField#behavior` method, which returned markup, was being masked
by the class's `attr_reader :behavior` definition, which returned the
attribute value passed in by the caller, resulting in markup similar to
this

    <input name="something" email>

instead of

    <input name="something" type="email">

While not broken, this resulted in incorrect behavior on the client.

This change moves the markup into the ERB layer, removing the clashing
method definition entirely.