feat(select): add support for `no-label`
Aligns the select component with the MDC web example markup when no
`label` DSL component is present in the `select` component.

Previously, when not specifying a label for a `select`, the generated
markup would result in an "empty" MDC outline notch where the label text
(an empty string) should be. This change adds the `mdc-select--no-label`
class to the component's element and only renders the `floating-label`
MDC element if the user has explicitly provided a `label` in the DSL.