fix: eliminate reload slowness/deadlock issue
Wrap the anonymous module returned by
`Rails.application.routes.url_helpers` in a `Singleton`. This ensures
that the route tree is not reconstructed every time a POM/helper calls a
route helper method defined by Rails.

This only affects development, as in non-development environments, files
never change and are thus not reloaded by the COPRL file watcher.
Outside of development, the `url_Helpers` module returned by Rails is
included as normal, NOT wrapped in a `Singleton`.