feat(grid): expand grid/column alignment
Grid/column alignment has been split into two properties: `align` and
`justify`.

`align` controls the layout of grid columns or column content along
the horizontal axis (for LTR and RTL layouts) and `justify` controls
the vertical axis.

Notably, this is a deliberate departure from the CSS box alignment
specification for grids, which behaves the opposite.

`:left` and `:right` grid/column alignment have been deprecated and
replaced with `:start`, `:center`, `:end`, and `:stretch` to better
support non-LTR layouts. To maintain backward compatibility, `:left` and
`:right` are mapped to `:start` and `:end`, respectively, and a
deprecation warning is emitted via the logger.

For example, this POM code

    grid do
      column align: :end, justify: :center do
      end
    end

results in a column with (for a left-to-right layout) right-aligned and
vertically centered content. For a user with a right-to-left layout,
this column's contents would be left-aligned.

The default values for `align` and `justify` are introduced as
`:stretch` to maintain backward compatibility.

Introduce grid/column `direction` attribute: `:row` or `:column`. This
attribute controls the direction of flow for a grid's columns or a
column's contents.

The default column direction value is `:row` to maintain backward
compatibility.