feat: add `parallel` event action
The `parallel` action allows users to dictate that a block of actions
are initiated simultaneously instead of sequentially. This is useful if
a more reactive UI is required and the replaced components are
potentially time-consuming to render.

`parallel` supports any action a normal event block would (including
loaded plugins).

For example, the following code will wait for the POST to `foos_path`
to succeed, begin replacing both `:some_stateful_component` and
`:something_else` at the same time, and finally (iff both `replace`
calls succeed) display a snackbar.

    event :click do
      post foos_path

      parallel do
        replace :some_stateful_component, 'a:presenter'
        replace :something_else, 'another:presenter'
      end

      snackbar 'did some work!'
    end