fix(events): use fragment to render replaced nodes
When replacing elements, use a DocumentFragment instead of a template
element to render the returned HTML as a node list.

Per the HTML spec, "script elements inserted using innerHTML do not
execute when they are inserted." This means that the previous usage of
setting `template.innerHTML` to "hydrate" replaced nodes is
fundamentally incompatible with the script COPRL plugin.

For more info, consult the HTML spec:
https://www.w3.org/TR/2008/WD-html5-20080610/dom.html#innerhtml0