fix(button): Fix buttons in nested form > dialog
In the web client, when a dialog with a raised button is nested within a
parent form (so: form > dialog > button), the button is rendered as
`type="submit"`. This inadvertently attempts to submit the parent form
instead of the dialog.

This fix causes raised buttons to be rendered as `type="submit'` only if
they're in a form. Otherwise, they're rendered as `type="button"`.