feat(dialog): Add option to prevent closing dialog
Adds the `dismissable` parameter to the `dialog` DSL action. When false,
the target dialog will not able to be closed via pressing the Escape key
nor by clicking/tapping the scrim after it has been opened.

For example:

    button 'Show dialog' do
      event :click do
        dialog :some_dialog, dismissable: false
      end
    end

The parameter defaults to true as a sane default and for backward
compatibility.