fix(web_client): Fix ERB inspection of parents
The web client's API application loses POM hierarchy when inflating JSON
to POM. This causes components which rely on checking their parent (or
an ancestor) for an attribute to exhibit different behavior and render
differently than the web client's view-rendering (non-API) application.

This change allows templates to pass down local variables (as one
expects to do with ERB), and updates the places relying on the above
behavior to use locals instead.