feat: Add support for "error group" elements
An error group element allows users to provide a closer target for
errors that would otherwise be rendered at the top of the page.

If a validation error refers to a nested field (e.g. `address[ln1]`),
and an `data-error-group="address"` element is on the page, any messages
for address fields will instead be rendered within the error group
element.

This logic applies only to page-level error message display. If a form
field for an error message is accessible, an error group matching the
message (if present) is not used.