feat: Add `appends` DSL action
The "appends" DSL action inserts new nodes to the tail of a target node:

    content id: :some_list

    button 'Append' do
      event :click do
        appends :some_list, :foo, some_param: 42
      end
    end

As with other DSL actions, `appends` is aliased as `append`.

In the web client, nodes are created from the HTML returned by fetching
a presenter. These nodes are inserted before the end of the target node,
then initialized.